+++ Changelog.md
@@ -5,6 +5,7 @@ Features:
  * Inline Assembly: Show useful error message if trying to access calldata variables.
  * Inline Assembly: Support variable declaration without initial value (defaults to 0).
  * Metadata: Only include files which were used to compile the given contract.
+ * Optimizer: Add new optimization step to remove unused ``JUMPDEST``s.
  * Type Checker: Disallow value transfers to contracts without a payable fallback function.
  * Type Checker: Include types in explicit conversion error message.
  * Type Checker: Raise proper error for arrays too large for ABI encoding.